Improve type safety of log interfaces.
