+++ projects/Apps/common/package.json
@@ -13,7 +13,9 @@
         "typescript": "^3.5.3",
         "@types/jest": "^24.0.17",
         "jest": "^25.1.0",
-        "ts-jest": "^25.1.0"
+        "ts-jest": "^25.1.0",
+        "@react-native-community/netinfo": "^5.6.2",
+        "@sentry/react-native": "^1.3.3"
     },
     "resolutions": {
         "jsdom": "16.2.1"
