+++ projects/Apps/common/src/cas-expiry.ts
@@ -1,0 +1,7 @@
+export interface CASExpiry {
+    content: string
+    expiryDate: string
+    expiryType: string
+    provider: string
+    subscriptionCode: string
+}
