+++ projects/Mallard/src/helpers/storage.ts
@@ -5,7 +5,7 @@ import {
     LEGACY_SUBSCRIBER_ID_USER_DEFAULT_KEY,
     LEGACY_SUBSCRIBER_POSTCODE_USER_DEFAULT_KEY,
 } from 'src/constants'
-import { CasExpiry } from 'src/services/content-auth-service'
+import { CASExpiry } from '../../../Apps/common/src/cas-expiry'
 import { ReceiptIOS } from 'src/authentication/services/iap'
 import { PushNotificationRegistration } from 'src/helpers/push-notifications'
 import { IdentityAuthData } from 'src/authentication/authorizers/IdentityAuthorizer'
@@ -54,7 +54,7 @@ const createAsyncCache = <T extends object | string | number>(key: string) => ({
     reset: (): Promise<void> => AsyncStorage.removeItem(key),
 })
 
-const casDataCache = createAsyncCache<CasExpiry>('cas-data-cache')
+const casDataCache = createAsyncCache<CASExpiry>('cas-data-cache')
 
 const userDataCache = createAsyncCache<IdentityAuthData>('user-data-cache')
